Move migrations to separate classes
